requirements.txt
fastapi
authx
python-dotenv